+++ processor/src/test/resources/input/com/example/inheritance/CompleteModel.java
@@ -1,0 +1,10 @@
+package com.example.inheritance;
+
+import arez.annotations.ArezComponent;
+import com.example.inheritance.other.BaseCompleteModel;
+
+@ArezComponent
+abstract class CompleteModel
+  extends BaseCompleteModel
+{
+}